+++ starsky/starskytest/FakeMocks/FakeIStorage.cs
@@ -273,6 +273,11 @@ namespace starskytest.FakeMocks
 
 		public bool WriteStream(Stream stream, string path)
 		{
+			if ( path == null )
+			{
+				throw new ArgumentNullException(nameof(path));
+			}
+			
 			_outputSubPathFiles.Add(path);
 
 			stream.Seek(0, SeekOrigin.Begin);